tutorial/pyproject.toml
[project]
name = "book-tutorial"
version = "0.1.0"
description = "Tutorial code for the book: Agents & RAG with LangGraph + LangChain"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "著者名（仮）" }]

dependencies = [
  "langgraph",
  "langchain",
  "langchain-openai",
  "langchain-anthropic",
  "pydantic>=2",
  "python-dotenv",
]

[dependency-groups]
dev = [
  "pytest",
  "ruff",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
addopts = "-q"
